nothing happenned is reflected in use-identified items list; troll agro radius is 1 bc too OP

+++ src/states/info-state.cpp
@@ -186,7 +186,7 @@ void InfoState::Update(bool forceUpdate)
 
 void InfoState::DrawScrollBars()
 {
-  if (_useIdentifiedMapCopy.size() > _th)
+  if (_useIdentifiedMapCopy.size() > (size_t)_th)
   {
     if (_scrollIndex == 0)
     {
